+++ src/main/java/com/roommake/admin/AdminController.java
@@ -8,89 +8,40 @@ import org.springframework.web.bind.annotation.RequestMapping;
 @RequestMapping("/admin")
 public class AdminController {
 
-	@GetMapping("/home")
-	public String adminHome() {
-		return "admin/home";
-	}
+    @GetMapping("/home")
+    public String adminHome() {
+        return "admin/home";
+    }
 
-	@GetMapping("/management")
-	public String notice() {
-		return "admin/management/notice";
-	}
+    @GetMapping("/management")
+    public String notice() {
+        return "admin/management/notice";
+    }
 
-	@GetMapping("/sales")
-	public String sales() {
-		return "admin/sales/sales";
-	}
+    @GetMapping("/sales")
+    public String sales() {
+        return "admin/sales/sales";
+    }
 
-	// 상품리스트
-	@GetMapping("/product/list")
-	public String product() {
-		return "admin/product/list";
-	}
-	// 상픔등록
-	@GetMapping("/product/insert")
-	public String insert(){
-		return "admin/product/insert";
-	}
-	//상품수정폼
-	@GetMapping("/product/modify")
-	public String modify(){
-		return "admin/product/modify";
-	}
+    // 상품리스트
+    @GetMapping("/product/list")
+    public String product() {
+        return "admin/product/list";
+    }
 
-	//상품 상세정보
-	@GetMapping("/product/detail")
-	public String detail() {
-		return "admin/product/detail";
-	}
-	//주문내역 리스트
-	@GetMapping("/order/item")
-	public String order() {
-		return "admin/order/item";
-	}
-	// 환불리스트
-	@GetMapping("order/refund")
-	public String refund(){
-		return "admin/order/refund";
-	}
+    //주문내역 리스트
+    @GetMapping("/order/item")
+    public String order() {
+        return "admin/order/item";
+    }
 
-	//교환페이지
-	@GetMapping("order/exchange")
-	public String exchange(){
-		return "admin/order/exchange";
-	}
+    @GetMapping("/user")
+    public String user() {
+        return "admin/user/user";
+    }
 
-	//교환 상세페이지
-	@GetMapping("/order/exchange_detail")
-	public String exchange_detail(){
-		return "admin/order/exchange_detail";
-	}
-	//반품페이지
-	@GetMapping("/order/return")
-	public String Return(){
-		return "admin/order/return";
-	}
-
-	//배송관리
-	@GetMapping("/order/status")
-	public String status(){
-		return "admin/order/status";
-	}
-
-	@GetMapping("order/delivery")
-	public String delivery(){
-		return "admin/order/delivery";
-	}
-
-	@GetMapping("/user")
-	public String user() {
-		return "admin/user/user";
-	}
-
-	@GetMapping("/community")
-	public String community() {
-		return "admin/community/community";
-	}
-
+    @GetMapping("/community")
+    public String community() {
+        return "admin/community/community";
+    }
 }
