+++ src/main/resources/templates/admin/order/exchange.html
@@ -1,4 +1,4 @@
-<html xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout" layout:decorate="~{layout/base-admin}"  th:with="menu='management'">
+<html xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout" layout:decorate="~{layout/base-admin}" th:with="menu='order'">
 <div layout:fragment="content-admin" class="container my-3">
     <div class="row mb-3">
         <div class="col-2">
@@ -8,48 +8,48 @@
             <h1>상품교환 리스트</h1>
             <table class="table">
                 <thead>
-                    <tr>
-                        <th>교환번호</th>
-                        <th>교환상세번호</th>
-                        <th>주문번호</th>
-                        <th>주문자명</th>
-                        <th>상품이름</th>
-                        <th>교환일자</th>
-                        <th>교환사유</th>
-                        <th>교환승인여부</th>
-                    </tr>
+                <tr>
+                    <th>교환번호</th>
+                    <th>교환상세번호</th>
+                    <th>주문번호</th>
+                    <th>주문자명</th>
+                    <th>상품이름</th>
+                    <th>교환일자</th>
+                    <th>교환사유</th>
+                    <th>교환승인여부</th>
+                </tr>
                 </thead>
                 <tbody>
-                    <tr>
-                        <td>50000001</td>
-                        <td>10515</td>
-                        <td><a class="link-offset-1" th:href="@{/admin/order/exchange_detail}">50628594</a></td>
-                        <td>이순신</td>
-                        <td>쇼파</td>
-                        <td>2022-05-05</td>
-                        <td>"쇼파 사이즈에 너무 작습니다"</td>
-                        <td align="center">교환신청</td>
-                    </tr>
-                    <tr>
-                        <td>50000002</td>
-                        <td>10515</td>
-                        <td><a class="link-offset-1" th:href="@{/admin/order/exchange_detail}">50628594</a></td>
-                        <td>홍길동</td>
-                        <td>쇼파</td>
-                        <td>2022-05-05</td>
-                        <td>"쇼파 사이즈에 너무 작습니다"</td>
-                        <td align="center">교환신청</td>
-                    </tr>
-                    <tr>
-                        <td>50000003</td>
-                        <td>10515</td>
-                        <td><a class="link-offset-1" th:href="@{/admin/order/exchange_detail}">50628594</a></td>
-                        <td>강감찬</td>
-                        <td>쇼파</td>
-                        <td>2022-05-05</td>
-                        <td>"쇼파 사이즈에 너무 작습니다"</td>
-                        <td align="center">교환신청</td>
-                    </tr>
+                <tr>
+                    <td>50000001</td>
+                    <td>10515</td>
+                    <td><a class="link-offset-1" th:href="@{/admin/order/exchange-detail}">50628594</a></td>
+                    <td>이순신</td>
+                    <td>쇼파</td>
+                    <td>2022-05-05</td>
+                    <td>"쇼파 사이즈에 너무 작습니다"</td>
+                    <td align="center">교환신청</td>
+                </tr>
+                <tr>
+                    <td>50000002</td>
+                    <td>10515</td>
+                    <td><a class="link-offset-1" th:href="@{/admin/order/exchange-detail}">50628594</a></td>
+                    <td>홍길동</td>
+                    <td>쇼파</td>
+                    <td>2022-05-05</td>
+                    <td>"쇼파 사이즈에 너무 작습니다"</td>
+                    <td align="center">교환신청</td>
+                </tr>
+                <tr>
+                    <td>50000003</td>
+                    <td>10515</td>
+                    <td><a class="link-offset-1" th:href="@{/admin/order/exchange-detail}">50628594</a></td>
+                    <td>강감찬</td>
+                    <td>쇼파</td>
+                    <td>2022-05-05</td>
+                    <td>"쇼파 사이즈에 너무 작습니다"</td>
+                    <td align="center">교환신청</td>
+                </tr>
                 </tbody>
             </table>
         </div>
