+++ src/main/resources/templates/admin/order/item.html
@@ -28,6 +28,7 @@
                     <th>주문수량</th>
                     <th>상품금액</th>
                     <th>배송상태</th>
+                    <th></th>
                     <th>송장번호</th>
                     <th>수령자</th>
                     <th>취소/반품여부</th>
@@ -43,39 +44,16 @@
                     <td>2개</td>
                     <td>100,000</td>
                     <td>
-                        <a th:href="@{/admin/order/status}">상품배송중</a>
+                        <select name="">
+                            <option value="">배송준비중</option>
+                            <option value="">배송중</option>
+                            <option value="">배송완료</option>
+                        </select>
                     </td>
-                    <td>15151-114828</td>
-                    <td>홍길동</td>
-                    <td align="center">N</td>
-                </tr>
-                <tr>
-                    <td>23</td>
-                    <td align="center">150</td>
-                    <td>2024-01-04</td>
-                    <td>4646-464</td>
-                    <td>책상</td>
-                    <td>2개</td>
-                    <td>100,000</td>
                     <td>
-                        <a th:href="@{/admin/order/status}">상품배송중</a>
+                        <button class="btn btn-primary btn-sm" type="submit">저장</button>
                     </td>
-                    <td>15151-114828</td>
-                    <td>홍길동</td>
-                    <td align="center">N</td>
-                </tr>
-                <tr>
-                    <td>23</td>
-                    <td align="center">150</td>
-                    <td>2024-01-04</td>
-                    <td>4646-464</td>
-                    <td>책상</td>
-                    <td>2개</td>
-                    <td>100,000</td>
-                    <td>
-                        <a th:href="@{/admin/order/status}">상품배송중</a>
-                    </td>
-                    <td>15151-114828</td>
+                    <td>151-118</td>
                     <td>홍길동</td>
                     <td align="center">Y</td>
                 </tr>
