+++ src/main/resources/templates/admin/order/refund.html
@@ -1,4 +1,4 @@
-<html xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout" layout:decorate="~{layout/base-admin}"  th:with="menu='management'">
+<html xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout" layout:decorate="~{layout/base-admin}" th:with="menu='order'">
 <div layout:fragment="content-admin" class="container my-3">
     <div class="row mb-3">
         <div class="col-2">
@@ -22,18 +22,18 @@
                                 <div class="col-10">
                                     <table class="table">
                                         <thead>
-                                            <tr>
-                                                <th>환불상품명</th>
-                                                <td>침대</td>
-                                                <th>환불포인트</th>
-                                                <td>10만포인트</td>
-                                            </tr>
-                                            <tr>
-                                                <th>신청자명</th>
-                                                <td>김유신</td>
-                                                <th>환불사유</th>
-                                                <td>침대 상태가 너무 않좋아요</td>
-                                            </tr>
+                                        <tr>
+                                            <th>환불상품명</th>
+                                            <td>침대</td>
+                                            <th>환불포인트</th>
+                                            <td>10만포인트</td>
+                                        </tr>
+                                        <tr>
+                                            <th>신청자명</th>
+                                            <td>김유신</td>
+                                            <th>환불사유</th>
+                                            <td>침대 상태가 너무 않좋아요</td>
+                                        </tr>
                                         </thead>
                                     </table>
                                 </div>
@@ -48,54 +48,54 @@
             </div>
             <table class="table">
                 <thead>
-                    <tr>
-                        <th></th>
-                        <th>환불번호</th>
-                        <th>주문번호</th>
-                        <th>상품이름</th>
-                        <th>환불금액</th>
-                        <th>환불접수일자</th>
-                        <th>환불포인트</th>
-                        <th>환불처리상태</th>
-                    </tr>
+                <tr>
+                    <th></th>
+                    <th>환불번호</th>
+                    <th>주문번호</th>
+                    <th>상품이름</th>
+                    <th>환불금액</th>
+                    <th>환불접수일자</th>
+                    <th>환불포인트</th>
+                    <th>환불처리상태</th>
+                </tr>
                 </thead>
                 <tbody>
-                    <tr>
-                        <th>
-                            <input  class="form-check-input" type="checkbox" value="">
-                        </th>
-                        <th>48448-4942</th>
-                        <td>48488</td>
-                        <td>쇼파</td>
-                        <td>50,000</td>
-                        <td>2024-04-16</td>
-                        <td>50,000</td>
-                        <td>환불신청</td>
-                    </tr>
-                    <tr>
-                        <th>
-                            <input  class="form-check-input" type="checkbox" value="">
-                        </th>
-                        <th>48448-4942</th>
-                        <td>48488</td>
-                        <td>쇼파</td>
-                        <td>50,000</td>
-                        <td>2024-04-16</td>
-                        <td>50,000</td>
-                        <td>환불신청</td>
-                    </tr>
-                    <tr>
-                        <th>
-                            <input class="form-check-input" type="checkbox" value="">
-                        </th>
-                        <th>5252-5252</th>
-                        <td>48488</td>
-                        <td>쇼파</td>
-                        <td>50,000</td>
-                        <td>2024-04-16</td>
-                        <td>50,000</td>
-                        <td>환불신청</td>
-                    </tr>
+                <tr>
+                    <th>
+                        <input class="form-check-input" type="checkbox" value="">
+                    </th>
+                    <th>48448-4942</th>
+                    <td>48488</td>
+                    <td>쇼파</td>
+                    <td>50,000</td>
+                    <td>2024-04-16</td>
+                    <td>50,000</td>
+                    <td>환불신청</td>
+                </tr>
+                <tr>
+                    <th>
+                        <input class="form-check-input" type="checkbox" value="">
+                    </th>
+                    <th>48448-4942</th>
+                    <td>48488</td>
+                    <td>쇼파</td>
+                    <td>50,000</td>
+                    <td>2024-04-16</td>
+                    <td>50,000</td>
+                    <td>환불신청</td>
+                </tr>
+                <tr>
+                    <th>
+                        <input class="form-check-input" type="checkbox" value="">
+                    </th>
+                    <th>5252-5252</th>
+                    <td>48488</td>
+                    <td>쇼파</td>
+                    <td>50,000</td>
+                    <td>2024-04-16</td>
+                    <td>50,000</td>
+                    <td>환불신청</td>
+                </tr>
                 </tbody>
             </table>
         </div>
