+++ src/main/resources/templates/admin/order/return.html
@@ -1,4 +1,4 @@
-<html xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout" layout:decorate="~{layout/base-admin}"  th:with="menu='management'">
+<html xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout" layout:decorate="~{layout/base-admin}" th:with="menu='order'">
 <div layout:fragment="content-admin" class="container my-3">
     <div class="row mb-3">
         <div class="col-2">
@@ -21,18 +21,18 @@
                                 <div class="10">
                                     <table class="table">
                                         <thead>
-                                            <tr>
-                                                <th>반품상품명</th>
-                                                <td>쇼파</td>
-                                                <th>반품금액</th>
-                                                <td>50,000원</td>
-                                            </tr>
-                                            <tr>
-                                                <th>신청자명</th>
-                                                <td>강감찬</td>
-                                                <th>반품사유</th>
-                                                <td>"쇼파가 찢어져서왔어요"</td>
-                                            </tr>
+                                        <tr>
+                                            <th>반품상품명</th>
+                                            <td>쇼파</td>
+                                            <th>반품금액</th>
+                                            <td>50,000원</td>
+                                        </tr>
+                                        <tr>
+                                            <th>신청자명</th>
+                                            <td>강감찬</td>
+                                            <th>반품사유</th>
+                                            <td>"쇼파가 찢어져서왔어요"</td>
+                                        </tr>
                                         </thead>
                                     </table>
                                 </div>
@@ -47,54 +47,54 @@
             </div>
             <table class="table">
                 <thead>
-                    <tr>
-                        <th></th>
-                        <th>반품번호</th>
-                        <th>반품신청일자</th>
-                        <th>주문자명</th>
-                        <th>반품처리상태</th>
-                        <th>주문상세번호</th>
-                        <th>반품사유</th>
-                        <th>반품승인여부</th>
-                    </tr>
+                <tr>
+                    <th></th>
+                    <th>반품번호</th>
+                    <th>반품신청일자</th>
+                    <th>주문자명</th>
+                    <th>반품처리상태</th>
+                    <th>주문상세번호</th>
+                    <th>반품사유</th>
+                    <th>반품승인여부</th>
+                </tr>
                 </thead>
                 <tbody>
-                    <tr>
-                        <td>
-                            <input class="form-check-input" type="checkbox" value="">
-                        </td>
-                        <td>159-1515</td>
-                        <td>2022-05-05</td>
-                        <td>이순신</td>
-                        <td>반품처리중</td>
-                        <td>202404168439</td>
-                        <td>"상태가 안좋습니다"</td>
-                        <td>반품승인</td>
-                    </tr>
-                    <tr>
-                        <td>
-                            <input class="form-check-input" type="checkbox" value="">
-                        </td>
-                        <td>159-15856</td>
-                        <td>2022-05-05</td>
-                        <td>이순신</td>
-                        <td>반품처리중</td>
-                        <td>1616-499</td>
-                        <td>"상태가 안좋습니다"</td>
-                        <td>반품승인</td>
-                    </tr>
-                    <tr>
-                        <td>
-                            <input class="form-check-input" type="checkbox" value="">
-                        </td>
-                        <td>5252-5252</td>
-                        <td>2022-05-05</td>
-                        <td>이순신</td>
-                        <td>반품처리중</td>
-                        <td>1616-499</td>
-                        <td>"상태가 안좋습니다"</td>
-                        <td>반품승인</td>
-                    </tr>
+                <tr>
+                    <td>
+                        <input class="form-check-input" type="checkbox" value="">
+                    </td>
+                    <td>159-1515</td>
+                    <td>2022-05-05</td>
+                    <td>이순신</td>
+                    <td>반품처리중</td>
+                    <td>202404168439</td>
+                    <td>"상태가 안좋습니다"</td>
+                    <td>반품승인</td>
+                </tr>
+                <tr>
+                    <td>
+                        <input class="form-check-input" type="checkbox" value="">
+                    </td>
+                    <td>159-15856</td>
+                    <td>2022-05-05</td>
+                    <td>이순신</td>
+                    <td>반품처리중</td>
+                    <td>1616-499</td>
+                    <td>"상태가 안좋습니다"</td>
+                    <td>반품승인</td>
+                </tr>
+                <tr>
+                    <td>
+                        <input class="form-check-input" type="checkbox" value="">
+                    </td>
+                    <td>5252-5252</td>
+                    <td>2022-05-05</td>
+                    <td>이순신</td>
+                    <td>반품처리중</td>
+                    <td>1616-499</td>
+                    <td>"상태가 안좋습니다"</td>
+                    <td>반품승인</td>
+                </tr>
                 </tbody>
             </table>
         </div>
