--- src/main/resources/templates/admin/order/status.html
@@ -1,29 +1,0 @@
-<html xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout" layout:decorate="~{layout/base-admin}"  th:with="menu='management'">
-<div layout:fragment="content-admin" class="container my-3">
-    <div class="row mb-3">
-        <div class="col-2">
-            <div th:replace="~{layout/sidebar-admin :: sidebar-admin-fragment}"></div>
-        </div>
-        <div class="col-10">
-            <h1>주문상태</h1>
-            <table class="table table-bordered"></table>
-            <div class="form-check form-check-inline">
-                <input class="form-check-input" type="radio" name="inlineRadioOptions" id="inlineRadio1" value="option1">
-                <label class="form-check-label" for="inlineRadio1">상품준비중</label>
-            </div>
-            <div class="form-check form-check-inline">
-                <input class="form-check-input" type="radio" name="inlineRadioOptions" id="inlineRadio2" value="option2">
-                <label class="form-check-label" for="inlineRadio2">상품배송중</label>
-            </div>
-            <div class="form-check form-check-inline">
-                <input class="form-check-input" type="radio" name="inlineRadioOptions" id="inlineRadio3" value="option2">
-                <label class="form-check-label" for="inlineRadio2">상품배송완료</label>
-            </div>
-            <div class="d-grid gap-2 d-md-flex justify-content-md-end">
-                <a class="btn btn-primary" th:href="@{/admin/order}">저장</a>
-            </div>
-
-        </div>
-    </div>
-</div>
-</html>
