+++ src/main/resources/templates/admin/product/detail.html
@@ -1,8 +1,8 @@
-<html xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout" layout:decorate="~{layout/base-admin}"  th:with="menu='management'">
+<html xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout" layout:decorate="~{layout/base-admin}" th:with="menu='product'">
 <div layout:fragment="content-admin" class="container my-3">
     <div class="row mb-3">
         <div class="col-2">
-            <div  th:replace="~{layout/sidebar-admin :: sidebar-admin-fragment}"></div>
+            <div th:replace="~{layout/sidebar-admin :: sidebar-admin-fragment}"></div>
         </div>
         <div class="col-10">
             <div class="row mb-3">
@@ -13,30 +13,30 @@
                     <h1>제품상세설명</h1>
                     <table class="table">
                         <tbody>
-                            <tr>
-                                <th>상품번호</th>
-                                <td>56</td>
-                                <th>상품명</th>
-                                <td>쇼파</td>
-                            </tr>
-                            <tr>
-                                <th>카테고리</th>
-                                <td>가구,빨간색</td>
-                                <th>상품가격</th>
-                                <td>50,000원</td>
-                            </tr>
-                            <tr>
-                                <th>재고</th>
-                                <td>50개</td>
-                                <th>할인률</th>
-                                <td>30%</td>
-                            </tr>
-                            <tr>
-                                <th>상품등록일</th>
-                                <td>2024-01-01</td>
-                                <th>상품수정일</th>
-                                <td>2024-03-04</td>
-                            </tr>
+                        <tr>
+                            <th>상품번호</th>
+                            <td>56</td>
+                            <th>상품명</th>
+                            <td>쇼파</td>
+                        </tr>
+                        <tr>
+                            <th>카테고리</th>
+                            <td>가구,빨간색</td>
+                            <th>상품가격</th>
+                            <td>50,000원</td>
+                        </tr>
+                        <tr>
+                            <th>재고</th>
+                            <td>50개</td>
+                            <th>할인률</th>
+                            <td>30%</td>
+                        </tr>
+                        <tr>
+                            <th>상품등록일</th>
+                            <td>2024-01-01</td>
+                            <th>상품수정일</th>
+                            <td>2024-03-04</td>
+                        </tr>
                         </tbody>
                     </table>
                     <div class="d-grid gap-2 d-md-flex justify-content-md-end">
@@ -52,12 +52,13 @@
                                             <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
                                         </div>
                                         <div class="form-group mb-3">
-                                            <label class="form-label">상위카테고리</label>
+                                            <label class="form-label">사이즈</label>
                                             <select class="form-select" name="select">
-                                                <option value="">국자</option>
-                                                <option value="">수세미</option>
+                                                <option value="">S</option>
+                                                <option value="">M</option>
+                                                <option value="">L</option>
                                             </select>
-                                            <label class="form-label">하위카테고리</label>
+                                            <label class="form-label">색상</label>
                                             <select class="form-select" name="select">
                                                 <option value="">빨강</option>
                                                 <option value="">파랑</option>
