+++ src/main/resources/templates/admin/product/list.html
@@ -6,9 +6,6 @@
         </div>
         <div class="col-10">
             <h1>상품리스트</h1>
-            <div class="modal-dialog modal-sm">
-                tq
-            </div>
             <div class="d-grid gap-2 d-md-flex justify-content-md-end">
                 <input type="button" class="btn btn-outline-dark" value="엑셀양식다운로드" onclick="download();"/>
                 <input type="button" class="btn btn-outline-dark" value="엘셀일괄업로드"/>
