+++ src/main/resources/templates/layout/sidebar-admin.html
@@ -27,7 +27,10 @@
         </div>
         <ul class="list-group list-group-flush">
             <li class="list-group-item" style="list-style: none">
-                <a th:href="@{/admin/product/list}" style="text-decoration:none;color: black">상품</a>
+                <a th:href="@{/admin/product/list}" style="text-decoration:none;color: black">상품리스트</a>
+            </li>
+            <li class="list-group-item" style="list-style: none">
+                <a th:href="@{/admin/product/insert}" style="text-decoration:none;color: black">상품등록</a>
             </li>
         </ul>
     </div>
